+++ lib/src/in_app_browser.dart
@@ -418,13 +418,13 @@ class InAppBrowser {
 
   ///Event fired when the WebView received an HTTP authentication request. The default behavior is to cancel the request.
   ///
-  ///[challenge] contains data about host, port, protocol, realm, etc. as specified in the [HttpAuthChallenge].
+  ///[challenge] contains data about host, port, protocol, realm, etc. as specified in the [URLAuthenticationChallenge].
   ///
   ///**Official Android API**: https://developer.android.com/reference/android/webkit/WebViewClient#onReceivedHttpAuthRequest(android.webkit.WebView,%20android.webkit.HttpAuthHandler,%20java.lang.String,%20java.lang.String)
   ///
   ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wknavigationdelegate/1455638-webview
   Future<HttpAuthResponse?>? onReceivedHttpAuthRequest(
-      HttpAuthChallenge challenge) {}
+      URLAuthenticationChallenge challenge) {}
 
   ///Event fired when the WebView need to perform server trust authentication (certificate validation).
   ///The host application must return either [ServerTrustAuthResponse] instance with [ServerTrustAuthResponseAction.CANCEL] or [ServerTrustAuthResponseAction.PROCEED].
@@ -762,6 +762,16 @@ class InAppBrowser {
   ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wknavigationdelegate/1455643-webview
   Future<IOSNavigationResponseAction?>?
     iosOnNavigationResponse(IOSNavigationResponse navigationResponse) {}
+
+  ///Called when a web view asks whether to continue with a connection that uses a deprecated version of TLS (v1.0 and v1.1).
+  ///
+  ///[challenge] represents the authentication challenge.
+  ///
+  ///**NOTE**: available only on iOS 14.0+.
+  ///
+  ///**Official iOS API**: https://developer.apple.com/documentation/webkit/wknavigationdelegate/3601237-webview
+  Future<IOSShouldAllowDeprecatedTLSAction?>?
+    iosShouldAllowDeprecatedTLS(URLAuthenticationChallenge challenge) {}
 
   void throwIsAlreadyOpened({String message = ''}) {
     if (this.isOpened()) {